requirements.txt
absl-py==2.1.0
alabaster==0.7.12
altair==5.0.1
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.0
appscript==1.0.1
asgiref==3.2.10
asn1crypto==1.0.1
astroid==2.5.1
astropy==3.2.2
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.2.0
Automat==20.2.0
autopep8==1.5.5
awscli==1.18.58
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blinker==1.6.3
bokeh==1.3.4
boto==2.49.0
botocore==1.16.8
Bottleneck==1.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.8.4
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
degiroapi==0.9.5
distlib==0.3.1
distributed==2.5.2
distro==1.9.0
Django==3.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.0
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.1
flatbuffers==24.3.25
frozendict==2.4.0
fsspec==0.5.2
future==0.17.1
gast==0.3.3
gevent==1.4.0
gitdb==4.0.11
GitPython==3.1.42
glob2==0.7
gmpy2==2.0.8
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.16
greenlet==0.4.15
grpcio==1.30.0
gunicorn==20.0.4
h11==0.14.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==6.7.0
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.13.2
json5==0.8.5
jsonpickle==3.0.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
keras==2.11.0
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
libclang==18.1.1
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==5.1.0
macholib==1.11
Markdown==3.2.2
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.1.1
mccabe==0.6.1
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
multitasking==0.0.11
music21==7.3.3
mysql-connector-python==8.0.18
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
npTDMS==0.27.0
numba==0.45.1
numexpr==2.7.0
numpy==1.21.6
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
openai==1.12.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
openpyxl==3.0.0
opt-einsum==3.2.1
packaging==24.0
pandas==1.3.5
pandocfilters==1.4.2
parso==0.7.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
peewee==3.17.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
ppb==0.11.0
ppb-vector==1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.20.3
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
pycurl==7.43.0.3
pydantic==2.5.3
pydantic-core==2.14.6
pydeck==0.8.1b0
pydocstyle==5.1.1
pyflakes==2.2.0
pygame==2.0.1
pygments==2.17.2
PyHamcrest==2.0.2
PyInstaller==3.6
pylint==2.7.2
Pympler==1.0.1
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySDL2==0.9.7
pysdl2-dll==2.0.14.post1
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-dotenv==0.21.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-telegram==0.18.0
pytils==0.4.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
regex==2024.4.16
requests==2.31.0
requests-oauthlib==1.3.0
rich==13.7.1
rope==0.14.0
rsa==3.4.2
ruamel-yaml==0.15.46
s3transfer==0.3.3
safetensors==0.4.4
scikit-image==0.15.0
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
sqlparse==0.3.1
statsmodels==0.10.1
streamlit==1.23.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
telegram==0.0.1
telegram-text==0.1.2
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
tools==0.1.9
toolz==0.10.0
tornado==6.0.3
tqdm==4.66.5
traitlets==4.3.3
transformers==4.30.2
tweepy==3.10.0
Twisted==20.3.0
txppb==0.1
typed-ast==1.4.2
typing-extensions==4.7.1
tzdata==2024.1
tzlocal==4.3.1
ujson==4.0.2
unicodecsv==0.14.1
urllib3==1.24.2
validators==0.20.0
virtualenv==20.0.25
watchdog==3.0.0
wcwidth==0.1.7
webcolors==1.13
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.30.0
yfinance==0.2.37
zict==1.0.0
zipp==0.6.0
zope.interface==5.2.0
